# fix: table would still be rendered in JSON mode when writing to file