--- a/lib/validate.js
+++ b/lib/validate.js
@@ -66,6 +66,10 @@ module.exports = function validateOpts (opts, cbPassedIn) {
   // fill in defaults after
   opts = defaultOpts(opts)
 
+  if (opts.json === true) {
+    opts.renderProgressBar = opts.renderResultsTable = opts.renderLatencyTable = false
+  }
+
   if (opts.requests) {
     if (opts.requests.some(r => !isValidFn(r.setupRequest))) {
       return new Error('Invalid option setupRequest, please provide a function (or file path when in workers mode)')